Add files via upload
Inside the folders are the RES stuff like boss picture, maps, new tiles and objects.  The folders only have the new stuff to make it easier to copy and paste.

AssetSetter, CollisionChecker, MapConnections, TileManager, TileTypes are all the files that had code updated to fit the new maps and final boss. Don't forget the imports! 

Everything else can be dragged into objects as they are new. All the doors and the arrows I have been using to do map transitions